Navigation
Block: Block is enable move horizontal and restricted to move outside the grid, rotate and stop rotating and dropping.

GridManager has been added the assert of avaialbe move and rotate of the block.

main: Droping block can be navigated from main.

Scene: Some schange into the camera and and light. Add in a wall frame to visual display the restriction of the grid