Block dropping order after scoring
Landed blocks now is sorted by their minimum y pos on the grid.
It is to prevent for any previous blocks that has a greater y pos then the new blocks to take the position of the new blocks that falls because it will be rearrange it position before the scoring one.