Scoring section base part complete
The base (temporary) scoring code has been implanted and now seqvent for the function and flow are temporary implanted with that scoring with a whole row filled.

It can begin to the next step of the original scoring idea in planning by have a number of cube linked with each other to make it score.