Score factory pattern included and a New Bug discovered
A factory pattern for the base score to increase by its on is included into the game manager.
The base score of combo will be increasing by its own during the game play without preinserted from the start.

A new bug is during the testing when two cube with a sum is 8 will be scoring.
This is due to the grid itself has -1 when the cell is inoccupied by a cube.

The solution need to change the grid a bit by inserting cube object instead of have a negative value.