Bug 2 cube sum 8 scorering fixed
The abnormal scoring by 2 cube with sum of 8 can score had been solved.
It is due to the grid itself havs -1 on unregistrated cells.
Since the grid was to registrate a cube's value and use the grid itself to sum the value which led that two occupied cell which become 8 add one unoccupied cell -1 become 7.

The grid now have been altered that instead of a dictionary holding a list of ints, it'll be holding a list of cubes instead which the "null" can provide advances for checking the cell is unoccupied and during the summary of the cubes number.