Test the combo system
The combo system base is implanted and now the game scoring is currently switch back to the method of scoring with 3 align cube instead of 3 cubes value in sum to be 7.

For now the combo system seems to be working as it should be, and it'll be observe for futher improvments for the Gamemanager class.