GameOver condition development
Still working on the condition game play to make the game be game over if the row high reach the limit high.
Yet in the same time we are working on so the line can be adjustable for future use and other purpose.
The Blockmanager has been added with a method to return a check of if any of it storing blocks max high Y have surpassed the limit line and the GameManager class will be the adjuster for the line's high in the game.