Objective Factorised from GameManager
The Objective that was handled by the GameManager class has been factories to an own class with additional classes and improvement to handle the check and change the objective number.

Now the code with objective and achievements had been outcommented from the GameManager class in order to test the class.

In worse case it don't work, the project will back to the commit before THIS one.