Update scoringprocress moment
THis is an upgrade for the scoring progress method that had been in the requested feedback.

In order to display how the player scored from the last landed block and exactly how many combo had been made, I decide to focus to implament the upgrade of scoring calculation progress and display.

If it fails, return to the change before this push to the master.