Upgrade to scoring system
In order to avoid all those seperated fuction of "ScoreWithTwoCubes", "ScoreWithThreeCubes"..., I made a seperate class that load in add on vectors  dictionary of all possible combination.

With this the program can avoid of all those if section in each method with different combination manual, instead have a loop that iterate through those possible and available position combination.

The old system still remain until the testing of the new is working.
If not fetch the previous update before this.