Update: new rearrange in blockmanager
The floating blocks now has change its dropping method from loop to recrusive and instead of adding all the block that was dropping, now it'll only store those in the block list that first is vacant beneath them that will be checked right after the vanish animation is over.
For now everything is working as it should be, but still need to keep testing.
the bugfree version is on the commit 6bc44db.