Rewind update
The rewind system still is ongoing. It still off in many way, but it can be use if it return to the previous step.

The guideblock still cast in an error that is hard to detect, considering to change to have the block class have the prefab of the guide block