FloatPopup
The result from lab testing is working in lab scene, but in the mainGameScene is something off.
This is possible of because of the camera on the main game had been sat of because at first was the intent to let the lower left corner be the origo off the whole game.
A possible solvation now is to set the camera to (0,0) in order to have it center at the scene, move the table so its center point is at the origo instead and make all the block creates along with the offside vector that is about (2.25,4.5,0).