Bug fix discovered during popup developming
During the development with the scoring popup above the cubes in the game world, I have reset the camera back to (0,0) and make the blocks position be some offposition so the block will be landing on the correct position and the popup now is appearing correct above the scoring cubes.
During the development, I have discovered the code for positioning the block's subcube missplacing that was discovered and had been corrected.