Bug fix, block score not cube score
From before, it had been that once the block had manage to score, it containings cube don't score when it happened it linked with an another cube as 2 cube linking.

That is because the scoring position on the block happened to have the same scoring position that made the program been misgiven thought it the scoring was the same from earlier scoring data.
Now I have add in a Catigory ID to the scoring data if it is Block or Cube scoring.
For now it works and hopefully this was the cause for the other bug that the scoring don't work after the rearrangement after scoring progress of the original block.