TaskManager upgrade
The TaskManager have been moved into the Managers classes folder.

Some new function had been made and have "Task" with them that will replace the old that has "Objective"

A factorising had been made in GridManager that call RegistrateInfo that will be Replacing some repeating code in "ScoreWithCube" and "ScoreWithBlock" functions.
The code had been put to test and been proven that it registrate the same info like the code inside the loop in each method but had not been replaced yet.

It'll be in use once the TaskManager class finish the modification.