Bug Fix
It was a bug about when x1 and x5 task had the value 6, only the x1 task was scoring while x5 don't. The bug had been fixed by modifying the method TaskDataContainValue that now had been renamed as MatchTaskData will be checking the linkCube value along with the resultValue and return the rank.
The MatchTaskData will be replacing AchiveObjective.

A little modification had been done that the originalBlock will be moving back(forward in the direction in the code) in order to avoid collision with blocks that already had landed