Exponential decay dropspeed implented
The funciton for the dropping speed to decay with exponential formula is included in the new class MathTools.

Now the current droprate will be taken from the DropRate Property in GameManager class and it'll update it value when the levels increase.

MathToos is an assembly class that will assamble complicated math calculation formula that will be useful for the game.