First stage of upgrade input system
A input manager is now implanted into the project, and hand been testing use in the main game with a few inputs command.

In order to have a more systematical system for the game, I have plan on making a manager class to hold all the menupanels and have a stack or finanstate machine system to navigate through the menu panels.

If it don't work, go back to this commit