New Control system in development
I keep working on the control system by creating polymaphism classes for different control.

It still missing a small thing for the one I work on, but I load down the new input system from the packed manager but hasn't put it online yet.

Will be testing it after this commit, so if it doesn't work, restart from this commit