InputManager
The new input manager need time to experment so I will be keep building the inputmanager I'm working on.
For now the keyboard is working with the command, and to progress to xbox 360 gamepad, I will need to change the original inputManager in the unity editor to work.

If after the change of the IM is not working, come back to this commit