New GUI manager system
The new GUI manager system with navigation and input support by the new developed controlsystem "Control manager" is working on it prototype state.
In the future the GUI manager will keep using the MenuPanelBase as abstract class and MenuManager to handle the navigation between the GUI panels.
OBS, class name to MenuPanelBase and MenuManager might change in the future.