Everything done to now
All that had been done till now is the keybinding gui along with the new xbox control class and tested it works as it should in the game.

A new main game scene by dublicate from the original game scene has been added which will be serve as the developing scene for new gameplay and GUI components.

The original is still working and will be use for game play testing until the new scene is fully developed and working properly.